--- DropoutCoder.EntityFramework.SharedTransaction/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="EntityFramework" version="6.4.4" targetFramework="net481" />
-</packages>
