--- DropoutCoder.EntityFrameworkCore.SharedTransaction/Database/DatabaseContextFactory.cs
@@ -1,17 +1,0 @@
-﻿namespace DropoutCoder.EntityFrameworkCore.SharedTransaction.Database
-{
-    using DropoutCoder.EntityFramework.SharedTransaction.Database;
-    using Microsoft.EntityFrameworkCore;
-    using Microsoft.EntityFrameworkCore.Design;
-
-    public class DatabaseContextFactory : IDesignTimeDbContextFactory<DatabaseContext>
-    {
-        public DatabaseContext CreateDbContext(string[] args)
-        {
-            var optionsBuilder = new DbContextOptionsBuilder<DatabaseContext>()
-                .UseSqlServer(DatabaseContext.ConnectionString);
-
-            return new DatabaseContext(optionsBuilder.Options);
-        }
-    }
-}
