--- DropoutCoder.EntityFrameworkCore.SharedTransaction/Database/Schema/Entity.cs
@@ -1,9 +1,0 @@
-﻿namespace DropoutCoder.EntityFramework.SharedTransaction.Database.Schema
-{
-    public class Entity
-    {
-        public int Id { get; set; }
-
-        public string Name { get; set; }
-    }
-}
