--- DropoutCoder.EntityFrameworkCore.SharedTransaction/DropoutCoder.EntityFrameworkCore.SharedTransaction.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.10.34916.146
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DropoutCoder.EntityFrameworkCore.SharedTransaction", "DropoutCoder.EntityFrameworkCore.SharedTransaction.csproj", "{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}.Debug|x64.ActiveCfg = Debug|x64
-		{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}.Debug|x64.Build.0 = Debug|x64
-		{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}.Release|x64.ActiveCfg = Release|x64
-		{6E2FE2CD-4065-42D8-AEFC-4C18AED10856}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8712A77B-AAE9-4B79-82AA-442D5E978113}
-	EndGlobalSection
-EndGlobal
