+++ src/EntityFramework.SharedTransaction.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.34916.146
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EntityFramework.SharedTransaction", "EntityFramework.SharedTransaction.csproj", "{D7E34DBD-6478-41F8-8268-A70225D5BFC1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Debug|x64.ActiveCfg = Debug|x64
+		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Debug|x64.Build.0 = Debug|x64
+		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Release|x64.ActiveCfg = Release|x64
+		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7068312F-0BBC-4217-879E-A52D31E0BBA7}
+	EndGlobalSection
+EndGlobal
