+++ src/EntityFramework.SharedTransaction/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("DropoutCoder.EntityFramework.SharedTransaction")]
+[assembly: AssemblyTitle("EntityFramework.SharedTransaction")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("DropoutCoder.EntityFramework.SharedTransaction")]
+[assembly: AssemblyProduct("EntityFramework.SharedTransaction")]
 [assembly: AssemblyCopyright("Copyright ©  2024")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
