+++ src/EntityFramework.SharedTransaction/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EntityFramework" version="6.5.1" targetFramework="net481" />
+</packages>
