+++ src/EntityFrameworkCore.SharedTransaction/Database/DatabaseContextFactory.cs
@@ -1,0 +1,14 @@
+﻿namespace EntityFrameworkCore.SharedTransaction.Database; 
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Design;
+
+public class DatabaseContextFactory : IDesignTimeDbContextFactory<DatabaseContext>
+{
+    public DatabaseContext CreateDbContext(string[] args)
+    {
+        var optionsBuilder = new DbContextOptionsBuilder<DatabaseContext>()
+            .UseSqlServer(DatabaseContext.ConnectionString);
+
+        return new DatabaseContext(optionsBuilder.Options);
+    }
+}
