+++ src/EntityFrameworkCore.SharedTransaction/Database/Schema/Entity.cs
@@ -1,0 +1,7 @@
+﻿namespace EntityFrameworkCore.SharedTransaction.Database.Schema; 
+public class Entity
+{
+    public int Id { get; set; }
+
+    public string Name { get; set; }
+}
