+++ src/SharedTransaction.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.34916.146
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DropoutCoder.EntityFramework.SharedTransaction", "DropoutCoder.EntityFramework.SharedTransaction.csproj", "{D7E34DBD-6478-41F8-8268-A70225D5BFC1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EntityFramework.SharedTransaction", "EntityFramework.SharedTransaction\EntityFramework.SharedTransaction.csproj", "{D7E34DBD-6478-41F8-8268-A70225D5BFC1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EntityFrameworkCore.SharedTransaction", "EntityFrameworkCore.SharedTransaction\EntityFrameworkCore.SharedTransaction.csproj", "{EAB93D92-22A4-769B-F6CE-941AE81B764A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,14 @@ Global
 		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Release|x64.ActiveCfg = Release|x64
 		{D7E34DBD-6478-41F8-8268-A70225D5BFC1}.Release|x64.Build.0 = Release|x64
+		{EAB93D92-22A4-769B-F6CE-941AE81B764A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EAB93D92-22A4-769B-F6CE-941AE81B764A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EAB93D92-22A4-769B-F6CE-941AE81B764A}.Debug|x64.ActiveCfg = Debug|x64
+		{EAB93D92-22A4-769B-F6CE-941AE81B764A}.Debug|x64.Build.0 = Debug|x64
+		{EAB93D92-22A4-769B-F6CE-941AE81B764A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EAB93D92-22A4-769B-F6CE-941AE81B764A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EAB93D92-22A4-769B-F6CE-941AE81B764A}.Release|x64.ActiveCfg = Release|x64
+		{EAB93D92-22A4-769B-F6CE-941AE81B764A}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
